requirements.txt
cloudinary==1.15.0
coverage==4.5.2
coveralls==1.5.1
dj-database-url==0.5.0
Django==2.1.5
django-cloudinary-storage==0.2.3
django-cors-middleware==1.3.1
django-extensions==2.1.4
djangorestframework==3.9.0
django-mailer==1.2.5
gunicorn==19.9.0
drf-yasg==1.12.1
Pillow==5.4.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
PyJWT==1.7.1
python-decouple==3.1
pytz==2018.9
six==1.12.0
whitenoise==4.1.2
PyYAML==3.13